Refactor Hindley Milner
